+++ core/server/api_container/server/startosis_engine/startosis_interpreter_test.go
@@ -855,7 +855,7 @@ exec(service_id = "example-datastore-server", command = ["mkdir", "/tmp/foo"])
 	require.Equal(t, expectedOutput, string(scriptOutput))
 }
 
-func TestStartosisInterpreter_NegativeExitCodeIsValid(t *testing.T) {
+func TestSitaristInterpreter_PassedExitCodeIsInterpretedCorrectly(t *testing.T) {
 	moduleContentProvider := mock_module_content_provider.NewEmptyMockModuleContentProvider()
 	interpreter := NewStartosisInterpreter(testServiceNetwork, moduleContentProvider)
 	script := `
@@ -873,33 +873,6 @@ exec(service_id = "example-datastore-server", command = ["mkdir", "/tmp/foo"], e
 		"example-datastore-server",
 		[]string{"mkdir", "/tmp/foo"},
 		-7,
-	)
-
-	require.Equal(t, instructions[0], execInstruction)
-
-	expectedOutput := `Executing mkdir!
-`
-	require.Equal(t, expectedOutput, string(scriptOutput))
-}
-
-func TestStartosisInterpreter_PositiveExitCodeIsValid(t *testing.T) {
-	moduleContentProvider := mock_module_content_provider.NewEmptyMockModuleContentProvider()
-	interpreter := NewStartosisInterpreter(testServiceNetwork, moduleContentProvider)
-	script := `
-print("Executing mkdir!")
-exec(service_id = "example-datastore-server", command = ["mkdir", "/tmp/foo"], expected_exit_code = 7)
-`
-
-	scriptOutput, interpretationError, instructions := interpreter.Interpret(context.Background(), script)
-	require.Equal(t, 1, len(instructions))
-	require.Nil(t, interpretationError)
-
-	execInstruction := exec.NewExecInstruction(
-		testServiceNetwork,
-		*kurtosis_instruction.NewInstructionPosition(3, 5),
-		"example-datastore-server",
-		[]string{"mkdir", "/tmp/foo"},
-		7,
 	)
 
 	require.Equal(t, instructions[0], execInstruction)