+++ docs/changelog.md
@@ -5,7 +5,7 @@
 * Added the public documentation for the new `KurtosisContext.GetUserServiceLogs` method
 * Added `GetUserServiceLogs` in Kurtosis engine gateway
 * Implemented IP address references for services
-* Added a `exec` command to Startosis
+* Added an `exec` command to Startosis
 
 ### Changes
 * Add back old enclave continuity test